haku: rename the dashboard "arm" → "console"
Mechanical rename across the monorepo, per operator preference:

- Python package haku/arm → haku/console (module path haku.arm → haku.console,
  main_module, imports, tests); CONSOLE_SIGNATURE; FastAPI title.
- k8s Service/Deployment + flux-kustomization haku-arm → haku-console; root
  kustomization entry; Authentik internal_host → haku-console.haku-sandbox.
- GHCR image ghcr.io/agentydragon/haku-arm → haku-console; Flux
  ImageRepository/ImagePolicy + imagepolicy marker; webhook-receiver registration;
  push-images matrix row.
- HAKU_ARM_ env prefix → HAKU_CONSOLE_; git commit identity haku-arm → haku-console;
  commit-message prefix arm: → console:.
- Prose: instructions.md, run.md, PLAN.md, schema/item.json, README.

The service was never merged to devel, so the old GHCR image and Flux image objects
never existed — nothing to GC.
