+++ cluster/k8s/flux-image-automation-ghcr/haku-console-image-policy.yaml
@@ -1,11 +1,11 @@
 apiVersion: image.toolkit.fluxcd.io/v1beta2
 kind: ImagePolicy
 metadata:
-  name: haku-arm
+  name: haku-console
   namespace: flux-system
 spec:
   imageRepositoryRef:
-    name: haku-arm
+    name: haku-console
   filterTags:
     pattern: '^devel-\d{14}-[0-9a-f]{7}$'
   policy:
