+++ cluster/k8s/flux-image-automation-ghcr/haku-console-image-repository.yaml
@@ -1,8 +1,8 @@
 apiVersion: image.toolkit.fluxcd.io/v1beta2
 kind: ImageRepository
 metadata:
-  name: haku-arm
+  name: haku-console
   namespace: flux-system
 spec:
-  image: ghcr.io/agentydragon/haku-arm
+  image: ghcr.io/agentydragon/haku-console
   interval: 5m
