+++ cluster/k8s/haku/console/flux-kustomization.yaml
@@ -1,13 +1,13 @@
 apiVersion: kustomize.toolkit.fluxcd.io/v1
 kind: Kustomization
 metadata:
-  name: haku-arm
+  name: haku-console
   namespace: flux-system
 spec:
   interval: 10m
   retryInterval: 1m
   timeout: 5m
-  path: ./cluster/k8s/haku/arm
+  path: ./cluster/k8s/haku/console
   prune: true
   wait: true
   sourceRef:
