+++ cluster/k8s/haku/console/service.yaml
@@ -1,14 +1,14 @@
 apiVersion: v1
 kind: Service
 metadata:
-  name: haku-arm
+  name: haku-console
   namespace: haku-sandbox
   labels:
-    app.kubernetes.io/name: haku-arm
+    app.kubernetes.io/name: haku-console
 spec:
   type: ClusterIP
   selector:
-    app.kubernetes.io/name: haku-arm
+    app.kubernetes.io/name: haku-console
   ports:
     - name: http
       port: 8080
