+++ cluster/k8s/kustomization.yaml
@@ -201,7 +201,7 @@ resources:
   # --- haku (personal background agent; see haku/PLAN.md) ---
   - haku/namespace/flux-kustomization.yaml
   - haku/rbac/flux-kustomization.yaml
-  - haku/arm/flux-kustomization.yaml
+  - haku/console/flux-kustomization.yaml
 
   # --- matrix ---
   - matrix/namespace/flux-kustomization.yaml
