+++ haku/console/AGENTS.md
