+++ haku/console/templates/page.html.j2
