+++ haku/console/templates_loader.py
