+++ haku/console/test_app.py
@@ -1,8 +1,8 @@
 """Integration test: the FastAPI app over a seeded local haku-state clone.
 
 The write endpoints are exercised against a real (local, bare) git remote, so each
-test asserts the commit actually landed on the remote with the ``haku-arm`` identity
-— the arm's whole job is to turn an operator click into a git event Haku can reduce.
+test asserts the commit actually landed on the remote with the ``haku-console`` identity
+— the console's whole job is to turn an operator click into a git event Haku can reduce.
 """
 
 from __future__ import annotations
@@ -13,7 +13,7 @@ import pygit2
 import pytest_bazel
 from fastapi.testclient import TestClient
 
-from haku.arm.app import create_app
+from haku.console.app import create_app
 
 
 def _remote_tip(bare: Path) -> pygit2.Commit:
@@ -42,8 +42,8 @@ def test_click_records_overlay_and_renders_clicked(seeded) -> None:
         assert f'action="/items/{item_id}/actions/snooze/unclick"' in client.get("/").text
     assert (item_id, "snooze") in seeded.git_state.read_clicks()
     tip = _remote_tip(seeded.bare)
-    assert tip.author.name == "haku-arm"
-    assert tip.message == f"arm: click snooze on {item_id}"
+    assert tip.author.name == "haku-console"
+    assert tip.message == f"console: click snooze on {item_id}"
 
 
 def test_unclick_retracts_the_overlay(seeded) -> None:
@@ -56,7 +56,7 @@ def test_unclick_retracts_the_overlay(seeded) -> None:
         assert f'action="/items/{item_id}/actions/snooze"' in page  # back to a plain click
         assert "/unclick" not in page
     assert seeded.git_state.read_clicks() == set()
-    assert _remote_tip(seeded.bare).message == f"arm: unclick snooze on {item_id}"
+    assert _remote_tip(seeded.bare).message == f"console: unclick snooze on {item_id}"
 
 
 def test_delete_clears_the_click(seeded) -> None:
@@ -73,8 +73,8 @@ def test_feedback_appends_intake_note(seeded) -> None:
     with TestClient(app) as client:
         assert client.post("/feedback", data={"text": "please prioritize taxes"}).status_code == 200
     tip = _remote_tip(seeded.bare)
-    assert tip.author.name == "haku-arm"
-    assert tip.message == "arm: feedback"
+    assert tip.author.name == "haku-console"
+    assert tip.message == "console: feedback"
     notes = list((seeded.settings.clone_dir / "intake").glob("*-feedback.md"))
     assert len(notes) == 1
     assert "please prioritize taxes" in notes[0].read_text()
